+++ src/components/chevronous-text/examples/chevronous-text-example-basic.tsx
@@ -1,0 +1,13 @@
+/*
+Basic.
+*/
+import React from 'react';
+import ChevronousText from '../chevronous-text';
+
+export default function Example() {
+  return (
+    <div style={{ lineHeight: '24px' }}>
+      <ChevronousText text="Hello explorer" />
+    </div>
+  );
+}
